requirements.txt
einops
fvcore
libigl
matplotlib
numpy
pandas
Pillow
pillow-avif-plugin
plyfile
pyarrow
pytest
PyYAML
setuptools
tensorboard
torch
tqdm
trimesh
yacs
zipp
